Python Virtual Environment (venv)/Pv2/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.0.1
requests==2.32.4
urllib3==2.2.3
